feat: se generan 4 nuevos casos de verificacion y generacion de alertas y se separa logica de verificacion dentro de un modelo
